Seclude createTestService to dedicated integration tests utils
