+++ tests/integration-all/s3/tests.js
@@ -6,12 +6,8 @@ const { expect } = require('chai');
 
 const { getTmpDirPath } = require('../../utils/fs');
 const { createBucket, createAndRemoveInBucket, deleteBucket } = require('../../utils/s3');
-const {
-  createTestService,
-  deployService,
-  removeService,
-  waitForFunctionLogs,
-} = require('../../utils/misc');
+const { deployService, removeService, waitForFunctionLogs } = require('../../utils/misc');
+const { createTestService } = require('../../utils/integration');
 const { getMarkers } = require('../shared/utils');
 
 describe('AWS - S3 Integration Test', function() {
