+++ tests/integration-all/schedule/tests.js
@@ -4,12 +4,8 @@ const path = require('path');
 const { expect } = require('chai');
 
 const { getTmpDirPath } = require('../../utils/fs');
-const {
-  createTestService,
-  deployService,
-  removeService,
-  waitForFunctionLogs,
-} = require('../../utils/misc');
+const { deployService, removeService, waitForFunctionLogs } = require('../../utils/misc');
+const { createTestService } = require('../../utils/integration');
 const { getMarkers } = require('../shared/utils');
 
 describe('AWS - Schedule Integration Test', function() {
